ZF2015-09: Fixed entropy issue in word CAPTCHA
This patch fixes a potential entropy fixation vector with `Zend_Captcha_Word`.
Prior to the fix, when selecting letters for the CAPTCHA, `array_rand()` was
used, which does not use sufficient entropy during randomization. The patch
backports randomization routines from ZF2 in order to provide a more
cryptographically secure RNG.